I have changed my local database screen to MVVM
